Fix calculating token cost if model does not support cache tokens
Not all models support cache tokens and so don't have price data
for it. In that case, it was throwing a KeyError and causing all
the costs to be 0.0. Now we do try/catch around each individual
cost so that if we fail to get one, we can still look to get the
others.

Signed-off-by: Matthew Peveler <mpeveler@timescale.com>